% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-CompactPME.Rmd.
\name{expr}
\alias{expr}
\title{Another helper function for the sphere kernel}
\usage{
expr(z)
}
\arguments{
\item{z}{The scalar value of some inner product}
}
\description{
Another helper function for the sphere kernel
}