% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-CompactPME.Rmd.
\name{fibonaccisphere}
\alias{fibonaccisphere}
\title{Sampling points on fibonacci sphere. Adapted from CoolTools library. Taken from github to avoid loading unnecessary functions}
\usage{
fibonaccisphere(n = 1000, r = 1)
}
\arguments{
\item{n}{sample size}

\item{r}{radius}
}
\description{
Sampling points on fibonacci sphere. Adapted from CoolTools library. Taken from github to avoid loading unnecessary functions
}