% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-CompactPME.Rmd.
\name{genSurface}
\alias{genSurface}
\title{Helper for generating a surface}
\usage{
genSurface(f)
}
\arguments{
\item{f}{A spline function}
}
\value{
A \code{list} of evaluation matrices
}
\description{
Helper for generating a surface
}