% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-CompactPME.Rmd.
\name{iso_try}
\alias{iso_try}
\title{An isomap method to find the smallest k so that the graph is not fragmented}
\usage{
iso_try(X, k = 1, d = 1)
}
\arguments{
\item{X}{the NxD data matrix}

\item{k}{the starting value for the k parameter search}

\item{d}{The dimension we wish to get values for. Use 1 for the line, 2 for S1, 3 for S2}
}
\value{
A list consisting of the \code{isomap} result and the optimal \code{k} value.
}
\description{
An isomap method to find the smallest k so that the graph is not fragmented
}