% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-CompactPME.Rmd.
\name{krst}
\alias{krst}
\title{The Reproducing kernel for S1}
\usage{
krst(r, s, t)
}
\arguments{
\item{r}{degree}

\item{s}{eval location 1}

\item{t}{eval location 2}
}
\description{
The Reproducing kernel for S1
}