% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-CompactPME.Rmd.
\name{krt}
\alias{krt}
\title{Helper function for the kernel}
\usage{
krt(r, t)
}
\arguments{
\item{r}{degree of the kernel}

\item{t}{where we evaluate}
}
\description{
Helper function for the kernel
}