% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-CompactPME.Rmd.
\name{local_var_1d}
\alias{local_var_1d}
\title{Compute smoothed residuals using kernel smoothing
@param t projection indices
@param resid_sq residual values
@param grid_eval grid size for smoothing
@param bandwidth bandwidth value. NULL for automatic selection
@param kernel the kernel method for ksmooth
@return A list consisting of the evaluation grid, the fitted residuals, and the bandwidth.}
\usage{
local_var_1d(
  t,
  resid_sq,
  periodic = F,
  grid_eval = 100,
  bandwidth = NULL,
  kernel = "normal"
)
}
\description{
Compute smoothed residuals using kernel smoothing
@param t projection indices
@param resid_sq residual values
@param grid_eval grid size for smoothing
@param bandwidth bandwidth value. NULL for automatic selection
@param kernel the kernel method for ksmooth
@return A list consisting of the evaluation grid, the fitted residuals, and the bandwidth.
}