% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-CompactPME.Rmd.
\name{local_var_2d}
\alias{local_var_2d}
\title{Compute smoothed residuals using kernel smoothing, 2d CASE
@param Z projection indices (vectors in R3)
@param resid_sq residual values
@param eval_size grid size for smoothing
@param bw_grid interval size for searching for optimal bandwidth
@return A list consisting of the evaluation grid, the fitted residuals, and the bandwidth.}
\usage{
local_var_2d(Z, resid_sq, eval_size = 1000, bw_grid = 100)
}
\description{
Compute smoothed residuals using kernel smoothing, 2d CASE
@param Z projection indices (vectors in R3)
@param resid_sq residual values
@param eval_size grid size for smoothing
@param bw_grid interval size for searching for optimal bandwidth
@return A list consisting of the evaluation grid, the fitted residuals, and the bandwidth.
}