% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-CompactPME.Rmd.
\name{normalize}
\alias{normalize}
\title{Normalizing for the 0,1 interval}
\usage{
normalize(t)
}
\arguments{
\item{t}{vector of projection indices}
}
\value{
The normalized values
}
\description{
Normalizing for the 0,1 interval
}