% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-CompactPME.Rmd.
\name{pelletier_kernel}
\alias{pelletier_kernel}
\title{Implements the pelletier kernel using Gaussian kernel}
\usage{
pelletier_kernel(x_news, x_data, y_data, h)
}
\arguments{
\item{x_news}{New x values to evaluate at}

\item{x_data}{The x data (projection indices)}

\item{y_data}{The y data (residuals)}
}
\value{
Estimated y values at the \code{x_news} values.
}
\description{
Implements the pelletier kernel using Gaussian kernel
}
\references{
Pelletier, B. (2006). Non-parametric regression estimation on closed Riemannian manifolds. Journal of Nonparametric Statistics, 18(1), 57-67.
}