% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-CompactPME.Rmd.
\name{plot_1d}
\alias{plot_1d}
\title{Plotting helper that calls the right function for 1d case}
\usage{
plot_1d(fit, S)
}
\arguments{
\item{fit}{A CPME object}

\item{S}{The parameterization. A vector of numbers between 0 and 1 for interval and S1. kx3 vectors in R3 where k is the desired fidelity}
}
\value{
A plotly object
}
\description{
Plotting helper that calls the right function for 1d case
}