% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-CompactPME.Rmd.
\name{plot_2d}
\alias{plot_2d}
\title{Helper for plotting a 2d surface}
\usage{
plot_2d(fit, S)
}
\arguments{
\item{fit}{A CPME object}

\item{S}{a kx3 \code{matrix} consisting of k-many vectors in R3, denoting the evaluation points of the spline}
}
\value{
Returns a \code{list} containing plotly objects
}
\description{
Helper for plotting a 2d surface
}