% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-CompactPME.Rmd.
\name{pme_2d_periodic}
\alias{pme_2d_periodic}
\title{The spline fit for the 1d Interval}
\usage{
pme_2d_periodic(
  data,
  lambdas,
  min_iter = 1,
  max_iter = 10,
  optimize_lambda = F,
  err_tol = 0.01,
  parallel = F,
  cl = NULL,
  init_iso = F,
  k_init = 10,
  resid_smooth_size = 1000,
  bw_grid = 1000,
  pelletier_bw = 0.1,
  eps_adj = 0
)
}
\arguments{
\item{data}{an Nx3 data matrix where N is the number of observations}

\item{lambdas}{A numeric vector of lambda values where we will perform a spline fit on. Lambda values typically are in the range of 1e-12 to 1e-2 for interpolation and for limiting lambda form respectively.}

\item{min_iter}{the minimum number of total iterations of PME. Default is 1 for speed, but choosing a higher number is useful if you want to visualize the cost function.}

\item{max_iter}{the maximum number of iterations of PME. Prevents the algorithm from stalling.}

\item{optimize_lambda}{whether or not to perform the method of section 5, Tuning Parameter Selection. If true, additional values will be returned, being the coefficient of variation at each lambda, along with the mean residual and standard deviation residual used to calculate it.}

\item{err_tol}{error tolerance for convergence of the algorithm. Lower values result in longer runtimes, and more likely to hit the maximum iteration cap.}

\item{parallel}{whether or not to run the algorithm parallel. The algorithm will be parallelized with respect to lambda and is recommended if your lambdas vector contains many values. Not recommended for only a handful of lamda values unless your dataset is large.}

\item{k_init}{The starting value for a search of the ISOMAP nearest neighbor parameter k. Our method is to find the smallest such k such that the ISOMAP graph is connected, as detailed in the iso_try function. The default is 1 to find the smallest possible.}

\item{eps_adj}{epsilon factor to add to denominator of CV calculation to prevent division by 0}

\item{lvp_grid}{The number denoting the size of the grid used for kernel smoothing for the tuning parameter selection. If left as NULL, it will choose a grid of 10 times the sample size.
#' @param cl A a makeCluster object for parallelization. For most cases, should be left null so that a new cluster is made upon runtime.}

\item{approx}{whether or not to use the approximate spline method. We recommend to set this to be TRUE for very large lambda vectors and or very large sample sizes.}
}
\description{
The spline fit for the 1d Interval
}