% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-CompactPME.Rmd.
\name{project_circle}
\alias{project_circle}
\title{Circular projection, for initialization puprposes}
\usage{
project_circle(X)
}
\arguments{
\item{X}{A data \code{matrix} consisting of N rows and 2 columns}
}
\value{
A vector of N values between 0 and 1 denoting the circular projection index
}
\description{
Circular projection, for initialization puprposes
}