% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-CompactPME.Rmd.
\name{project_grid}
\alias{project_grid}
\title{Optimization over a grid. This is to find the best projection index on a 1d Manifold}
\usage{
project_grid(Xraw, res, gridSize = 1000)
}
\arguments{
\item{Xraw}{A data \code{matrix}}

\item{res}{A spline fit function}

\item{gridSize}{The number of entries in the partition}
}
\value{
A vector of projection indices onto the spline
}
\description{
Optimization over a grid. This is to find the best projection index on a 1d Manifold
}