% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-CompactPME.Rmd.
\name{project_optimize}
\alias{project_optimize}
\title{Find projection index using an optimization routine. Grid method is preferred for global solutions}
\usage{
project_optimize(Xraw, res, periodic = F)
}
\arguments{
\item{Xraw}{A data \code{matrix}}

\item{res}{A spline fit function}

\item{periodic}{Whether or not the spline is periodic. \code{T} for the circle, \code{F} for the interval}
}
\value{
The projection indices
}
\description{
Find projection index using an optimization routine. Grid method is preferred for global solutions
}