% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-CompactPME.Rmd.
\name{project_optimize2}
\alias{project_optimize2}
\title{Computing projection indices for the spherical data case}
\usage{
project_optimize2(X, f, init)
}
\arguments{
\item{X}{Nx3 data matrix, where N is the number of observations}

\item{f}{The function to optimize over. Usually a spline fit}

\item{init}{An N-vector consisting of initial values for the optimizing search}
}
\description{
Computing projection indices for the spherical data case
}