% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-CompactPME.Rmd.
\name{Rk}
\alias{Rk}
\title{The reproducing kernel for the interval}
\usage{
Rk(s, t)
}
\arguments{
\item{s}{eval location 1. Should be a vector of equal length to t.}

\item{t}{eval location 2. Should be a vector of equal length to s.}
}
\value{
The RK evaluation
}
\description{
The reproducing kernel for the interval
}