% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-CompactPME.Rmd.
\name{spline1d_interval}
\alias{spline1d_interval}
\title{The spline fit for the 1d Interval}
\usage{
spline1d_interval(X, t, lambda)
}
\arguments{
\item{X}{the NxD data matrix where N is the number of observations and D is either 2 or 3.}

\item{t}{the set of projection indices associated with the observations of X. Should be a vector of length N.}

\item{lambda}{the regularization parameter. Should be a real number greater than 0, typically in the magnitude of 1e-12 (interpolating) to 1e-2 (corresponds to lambda=infinity case in paper).}
}
\value{
Returns the spline fit as a function and alpha value as an attribute.
}
\description{
The spline fit for the 1d Interval
}