% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-CompactPME.Rmd.
\name{spline1d_interval_approx}
\alias{spline1d_interval_approx}
\title{The approximate spline fit for the 1d Interval}
\usage{
spline1d_interval_approx(X, t, lambda)
}
\arguments{
\item{X}{the NxD data matrix where N is the number of observations and D is either 2 or 3.}

\item{t}{the set of projection indices associated with the observations of X. Should be a vector of length N.}

\item{lambda}{the regularization parameter.}
}
\value{
The spline fit as a function and alpha value as an attribute.
}
\description{
The approximate spline fit for the 1d Interval
}