% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-CompactPME.Rmd.
\name{spline1d_periodic}
\alias{spline1d_periodic}
\title{Fit spline}
\usage{
spline1d_periodic(X, t, lambda)
}
\arguments{
\item{X}{data}

\item{t}{projection indices}

\item{lambda}{smoothing parameter}
}
\description{
Fit spline
}