% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-CompactPME.Rmd.
\name{spline1d_periodic_approx}
\alias{spline1d_periodic_approx}
\title{Approximate spline for S1}
\usage{
spline1d_periodic_approx(X, t, lambda, k = 60, rescale_lam = F)
}
\arguments{
\item{X}{NxD data matrix}

\item{t}{projection indices}

\item{lambda}{smoothing parameter value}

\item{k}{number of basis functions for approximation}

\item{rescale_lam}{whether or not to rescale lambda so that the lambda choices for the approximate method correspond roughly to the exact method. Setting to T tries to make the scale more similar to the exact method, so identical lambda across the methods yield similar results}
}
\description{
Approximate spline for S1
}