% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-CompactPME.Rmd.
\name{spline2d}
\alias{spline2d}
\title{Spline on the sphere}
\usage{
spline2d(data, Tt, lambda)
}
\arguments{
\item{data}{An Nx3 data matrix}

\item{Tt}{An Nx3 projection index matrix. Each observation (row) is a vector parameterizing the location on the sphere S2 as opposed to longitude and latitude.}

\item{lambda}{The smoothing parameter. Real number greater than 0}
}
\description{
Spline on the sphere
}