% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-CompactPME.Rmd.
\name{spline_error}
\alias{spline_error}
\title{Computes residuals. Also works for the 2d case}
\usage{
spline_error(Xdata, t, splineresult)
}
\arguments{
\item{Xdata}{NxD data matrix}

\item{t}{vector of projection indices}

\item{splineresult}{a spline fit as a function}
}
\value{
The mean of squared residuals
}
\description{
Computes residuals. Also works for the 2d case
}