% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-CompactPME.Rmd.
\name{var_het}
\alias{var_het}
\title{Computes variance heterogneity given the results of local_var_1d. Also used for the 2d case.}
\usage{
var_het(sigma2_hat, eps_adj = 1e-14)
}
\arguments{
\item{sigma2_hat}{vector of smoothed MSE values}
}
\value{
A list consisting of the mean, standard deviation, coefficient of variation, and spread.
}
\description{
Computes variance heterogneity given the results of local_var_1d. Also used for the 2d case.
}