% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-CompactPME.Rmd.
\name{d_circ_vec}
\alias{d_circ_vec}
\title{Circular distances between a single vector and every other vector in a matrix}
\usage{
d_circ_vec(x_new, x_data)
}
\arguments{
\item{x_new}{A vector denoting a new data point}

\item{x_data}{The dataset}
}
\value{
A vector of distances between the vector to each row of the matrix using a circular geodesic
}
\description{
Circular distances between a single vector and every other vector in a matrix
}